convert concerns to helpers...concerns are for models..my bad
