+++ app/controllers/students_controller.rb
@@ -1,22 +1,6 @@
 class StudentsController < ApplicationController
-  # before_action :auth_check
-  #
-  # def auth_check
-  #   if(logged_in_app? ==false)
-  #     if(controller_name != "app_security")
-  #       redirect_to app_firewall_path
-  #     end
-  #   end
-  # end
-  #
-  # def logged_in_app?
-  #   if(session["appauth"] == true)
-  #     return true
-  #   end
-  #   return false
-  # end
-
-  include AppSecurityConcern
+  include AppSecurityHelper
+  before_action :auth_check
 
   def index
     render "students/new"
